remove anonymous class
